refactored input.py; sorted import blocks (ruff)
