+++ libs/langchain/langchain/callbacks/stdout.py
@@ -2,8 +2,8 @@
 from typing import Any, Dict, List, Optional, Union
 
 from langchain.callbacks.base import BaseCallbackHandler
-from langchain.input import print_text
 from langchain.schema import AgentAction, AgentFinish, LLMResult
+from langchain.utils.input import print_text
 
 
 class StdOutCallbackHandler(BaseCallbackHandler):
