+++ libs/langchain/langchain/callbacks/tracers/stdout.py
@@ -3,7 +3,7 @@ from typing import Any, Callable, List
 
 from langchain.callbacks.tracers.base import BaseTracer
 from langchain.callbacks.tracers.schemas import Run
-from langchain.input import get_bolded_text, get_colored_text
+from langchain.utils.input import get_bolded_text, get_colored_text
 
 
 def try_json_stringify(obj: Any, fallback: str) -> str:
