+++ libs/langchain/langchain/utils/__init__.py
@@ -6,6 +6,12 @@ These functions do not depend on any other langchain modules.
 
 from langchain.utils.env import get_from_dict_or_env, get_from_env
 from langchain.utils.formatting import StrictFormatter, formatter
+from langchain.utils.input import (
+    get_bolded_text,
+    get_color_mapping,
+    get_colored_text,
+    print_text,
+)
 from langchain.utils.math import cosine_similarity, cosine_similarity_top_k
 from langchain.utils.strings import comma_list, stringify_dict, stringify_value
 from langchain.utils.utils import (
@@ -24,11 +30,15 @@ __all__ = [
     "cosine_similarity",
     "cosine_similarity_top_k",
     "formatter",
+    "get_bolded_text",
+    "get_color_mapping",
+    "get_colored_text",
     "get_from_dict_or_env",
     "get_from_env",
     "get_pydantic_field_names",
     "guard_import",
     "mock_now",
+    "print_text",
     "raise_for_status_with_text",
     "stringify_dict",
     "stringify_value",
